Extract components list function for Altium from BOM
